Add support for Amazon Alexa
The goal is to have API.AI do all the language processing, because Alexa
is not that good at doing so. This does mean that there will be a slightly
longer delay for Alexa users to get the information as compared to Google
users because they will have to hit API.AI and wait for the translation
before the local processing.

Sizing has been added, with conversion to Imperial units because the target
audience is Americans.